Refactoring

+++ src/tests/NotepadBasedCalculator.Core.Tests/DataParserTests.cs
@@ -18,7 +18,7 @@ namespace NotepadBasedCalculator.Core.Tests
             Assert.Single(parserResult.Lines[0].Data);
             Assert.True(parserResult.Lines[0].Data[0].Is(PredefinedTokenAndDataTypeNames.Numeric));
             Assert.Equal(PredefinedTokenAndDataTypeNames.SubDataTypeNames.Integer, parserResult.Lines[0].Data[0].Subtype);
-            Assert.Equal(output, ((IntegerData)parserResult.Lines[0].Data[0]).Value);
+            Assert.Equal(output, ((DecimalData)parserResult.Lines[0].Data[0]).Value);
         }
 
         [Theory]
